Refactor: Route generation
Moves the route generation logic to a separate module. This improves code organization and reduces the size of the Gui struct.

The generate_random_routes and generate_random_not_flown_aircraft_routes functions are moved to the new RouteGenerator struct. This struct also contains the data needed for route generation, such as the list of airports and the spatial index.

The Gui struct now has a route_generator field, which is used to generate routes. The generate_random_routes and generate_random_not_flown_aircraft_routes methods are removed from the Gui struct.

This change makes the code more modular and easier to maintain. It also reduces the size of the Gui struct, which makes it easier to work with.